Refactor Hallo.js integration into a separate package.
