+++ wagtail/wagtailadmin/static_src/wagtailadmin/js/page-editor.js
@@ -1,64 +1,5 @@
 'use strict';
 
-var halloPlugins = {
-    halloformat: {},
-    halloheadings: {formatBlocks: ['p', 'h2', 'h3', 'h4', 'h5']},
-    hallolists: {},
-    hallohr: {},
-    halloreundo: {},
-    hallowagtaillink: {},
-    hallorequireparagraphs: {}
-};
-
-function registerHalloPlugin(name, opts) {
-    halloPlugins[name] = (opts || {});
-}
-
-function makeRichTextEditable(id) {
-    var input = $('#' + id);
-    var richText = $('<div class="richtext"></div>').html(input.val());
-    richText.insertBefore(input);
-    input.hide();
-
-    var removeStylingPending = false;
-    function removeStyling() {
-        /* Strip the 'style' attribute from spans that have no other attributes.
-        (we don't remove the span entirely as that messes with the cursor position,
-        and spans will be removed anyway by our whitelisting)
-        */
-        $('span[style]', richText).filter(function() {
-            return this.attributes.length === 1;
-        }).removeAttr('style');
-        removeStylingPending = false;
-    }
-
-    var closestObj = input.closest('.object');
-
-    richText.hallo({
-        toolbar: 'halloToolbarFixed',
-        toolbarCssClass: (closestObj.hasClass('full')) ? 'full' : (closestObj.hasClass('stream-field')) ? 'stream-field' : '',
-        plugins: halloPlugins
-    }).bind('hallomodified', function(event, data) {
-        input.val(data.content);
-        if (!removeStylingPending) {
-            setTimeout(removeStyling, 100);
-            removeStylingPending = true;
-        }
-    }).bind('paste', function(event, data) {
-        setTimeout(removeStyling, 1);
-    });
-}
-
-function insertRichTextDeleteControl(elem) {
-    var a = $('<a class="icon icon-cross text-replace delete-control">Delete</a>');
-    $(elem).addClass('rich-text-deletable').prepend(a);
-    a.click(function() {
-        $(elem).fadeOut(function() {
-            $(elem).remove();
-        });
-    });
-}
-
 function initDateChooser(id) {
     if (window.dateTimePickerTranslations) {
         $('#' + id).datetimepicker({
@@ -363,6 +304,17 @@ function initCollapsibleBlocks() {
                 $li.removeClass('collapsed');
                 $fieldset.show('show');
             }
+        });
+    });
+}
+
+
+function insertRichTextDeleteControl(elem) {
+    var a = $('<a class="icon icon-cross text-replace delete-control">Delete</a>');
+    $(elem).addClass('rich-text-deletable').prepend(a);
+    a.click(function() {
+        $(elem).fadeOut(function() {
+            $(elem).remove();
         });
     });
 }
