+++ wagtail/wagtailadmin/templates/wagtailadmin/pages/_editor_js.html
@@ -12,17 +12,15 @@
 
 {% compress js %}
     <script src="{% static 'wagtailadmin/js/vendor/rangy-core.js' %}"></script>
-    <script src="{% static 'wagtailadmin/js/vendor/hallo.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/vendor/tag-it.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/expanding_formset.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/modal-workflow.js' %}"></script>
-    <script src="{% static 'wagtailadmin/js/hallo-plugins/hallo-wagtaillink.js' %}"></script>
-    <script src="{% static 'wagtailadmin/js/hallo-plugins/hallo-hr.js' %}"></script>
-    <script src="{% static 'wagtailadmin/js/hallo-plugins/hallo-requireparagraphs.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/page-editor.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/page-chooser.js' %}"></script>
     <script src="{% static 'admin/js/urlify.js' %}"></script>
     <script src="{% static 'wagtailadmin/js/privacy-switch.js' %}"></script>
+
+    {% include "wagtailrichtexteditor/pages/_richtexteditor_js.html" %}
 
     {% hook_output 'insert_editor_js' %}
 {% endcompress %}
