+++ wagtail/wagtailforms/static_src/wagtailforms/js/rich-text-editor.js
@@ -1,0 +1,3 @@
+$(function() {
+
+});
