+++ wagtail/wagtailhallo/static_src/wagtailrichtexteditor/js/hallo-plugins/hallo-wagtailembeds.js
