+++ wagtail/wagtailhallo/templates/wagtailrichtexteditor/pages/_richtexteditor_js.html
@@ -1,0 +1,15 @@
+{% load wagtailadmin_tags compress static %}
+
+{% comment %}
+    Javascript declarations to be included on the 'create page' and 'edit page' views
+{% endcomment %}
+
+{% compress js %}
+    <script src="{% static 'wagtailrichtexteditor/js/vendor/hallo.js' %}"></script>
+    <script src="{% static 'wagtailrichtexteditor/js/hallo-plugins/hallo-wagtaillink.js' %}"></script>
+    <script src="{% static 'wagtailrichtexteditor/js/hallo-plugins/hallo-hr.js' %}"></script>
+    <script src="{% static 'wagtailrichtexteditor/js/hallo-plugins/hallo-requireparagraphs.js' %}"></script>
+    <script src="{% static 'wagtailrichtexteditor/js/rich-text-editor.js' %}"></script>
+
+    {% hook_output 'insert_richtexteditor_js' %}
+{% endcompress %}
